trace-cmd library: New API for creating an output trace handle
There is lack of API to create an output trace handle to file with
given name, which is common use case. Existing flow is -  the library
user should create the file, open it and pass a file descriptor to the
library for creating the output handle. The proposed new API
	tracecmd_output_create()
has file name as input parameter. It creates a file with given name (or
truncates existing one) and create an output trace handle to that file.
The file name parameter is optional, if NULL is passed - the created
output trace handle will not be associated with a file.

Link: https://lore.kernel.org/linux-trace-devel/20211202121949.43084-3-tz.stoyanov@gmail.com

Suggested-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>