trace-cmd: Set default log to INFO and remove unnecessary newlines
Set the default logging of trace-cmd report to INFO. This will not affect
the tep functions, as tep will still use whatever its default is. But it
does allow for normal prints to be shown by default.

For example, the version of the file that is being read should be reported
by default as it use to be.

Also remove all the unnecessary newlines that are added to the
tracecmd_info/warning/critical() functions, as they supply the newlines.

Link: https://lore.kernel.org/linux-trace-devel/20220221181509.7259bb7d@rorschach.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>