trace-cmd: Fix regression of not reading older data files
I have some old trace.dat files that for some reason have the version
missing from its file, causing it to flag a "version 0". Because there's
now a check for reading a minimum version, it fails to read these files
when it use to read them just fine. If there's no version number in the
file or the minimum is below what it thinks it is, then just read it.

Link: https://lore.kernel.org/linux-trace-devel/20220221213314.18516bf0@yoga.local.home

Fixes: 6482f99a47ac ("trace-cmd library: Define trace file version 7")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>