trace-cmd record: Add --name option to name unnamed guests
When using the vsocket port number to connect to a guest agent (-A @3:823)
the file that is created is called "unknown". This is not very
informative, and if you have more than one guest, it can cause confusion.

Add a "--name" option that allows the user to name the guest that it
connects with. The --name option takes one parameter (the name to call the
guest), and must be after a -A option.

Link: https://lore.kernel.org/linux-trace-devel/20220314171942.5b53f9e8@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>