trace-cmd library: Fix tracecmd_read_at()
The compression logic set all the file_offsets for the cpu_data to zero to
just use absolute offsets. But this breaks tracecmd_read_at() that has a
requirement that every record in a handle has a unique offset. This
includes records on different CPUs.

Make the file_offset unique and start at a new page boundary from the
previous cpu_data. Use next_offset in the input handle to keep track of
the last stored cpu_data and set the next cpu_data file_offset to it.

Also slightly restructure the input handle structure to be a little more
compact.

Link: https://lore.kernel.org/linux-trace-devel/20220317165537.46d5685b@gandalf.local.home

Fixes: 5113584096350 ("trace-cmd library: Initialize CPU data decompression logic")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>