trace-cmd: Fix bash completions for non specified commands
For trace-cmd commands that are not specifically specified by the bash
completion script, allow to see files after the commands. There's a bug in
the test that returns if there are options, and does not jump to the file
listing if there is any option even if the user does not specify one, and
wants a file instead.

Also make a reference to $cur more consistent with its other references,
which is ${cur}

Link: https://lore.kernel.org/linux-trace-devel/20220323225016.3a3a7034@gandalf.local.home

Fixes: 431f1a03f59cb trace-cmd: Fix bash completion on directory names
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>