trace-cmd library: Fix trace-cmd convert to handle offset update
With changing the cpu_data[]->file_offset to be unique per CPU, the
convert code had a path that was still relying on it being zero.
Specifically, when read_zpage is false. This would decompress the data
into a temp file, and the offset would be from that temp file. Thus, using
the unique file_offset would cause it to index farther than the size of
the file and crash. Instead, subtract the cpu_data[].file_offset from the
offset used in the temp file.

Link: https://lore.kernel.org/linux-trace-devel/20220404111547.33b58f5c@gandalf.local.home

Fixes: add83e0c8b51 ("trace-cmd library: Fix tracecmd_read_at()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>