trace-cmd library: Do not read if cpu is greater than cpus
Fix tracecmd_read_data() to not read a cpu if it is greater than the
number of CPUs recorded.

As tracecmd_cpus() returns the number of CPUs found in the system which
can be greater than the number of CPUs stored in the buffer, it is likely
that a caller will pass in a CPU number that is greater than the stored
array of cpu_data.

In this case return NULL otherwise it will access beyond the cpu_data
array.

Link: https://lore.kernel.org/linux-trace-devel/20220404111812.2b3753df@gandalf.local.home

Fixes: 8724996990021 ("trace-cmd library: Read extended BUFFER option")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>