trace-cmd library: Do not free pthread items from agent
The agent creates a tsync to synchronize with the host as well as the host
synchronizing with the guest. The host uses pthread synchronization items
(mutex, cond and barrier) to synchronize its tracing threads as it needs
to retrieve the data that is produced from them.

While debugging the agent, the forked connection would always fail with a
SIGFPE, Arithmetic exception. This was caused by calling
pthread_barrier_destroy() against a barrier that never was initialized.

Since the pthread_mutex, pthread_cond, and pthread_barriers are only
initialized by the host, skip the freeing of them (which both the host and
guest call the same function to free: tracecmd_tsync_free()), if the
guest_id is zero.

Link: https://lore.kernel.org/linux-trace-devel/20220406115638.44303e6a@gandalf.local.home

Fixes: 3c2d7f97245f1 ("trace-cmd: Wait for first time sync before the trace")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>