trace-cmd: Unlock tsync mutex on thread exit
The logic for synchronising host and guest time stamps uses pthreads and
mutexes. The guest pthread locks a mutex on the beginning of each
synchronization iteration and uses it to wait on a conditional variable.
When the synchronization is done and the thread exists, this mutex
should be unlocked.

Link: https://lore.kernel.org/linux-trace-devel/20220411041002.132379-1-tz.stoyanov@gmail.com

fixes: 08b9d507 ("trace-cmd: Basic infrastructure for host - guest timestamp synchronization")
Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=215808
Reported-by: Ryan <ycaibb@gmail.com>
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>