trace-cmd: Write the actual size of compressed block in the trace file
Before a data block is compressed using the chosen compression
algorithm, a hook function is called to calculate the size of the
required buffer to compress the data. In most cases that size is bigger
than the actual size of the compressed data. The current logic writes
into the trace file that buffer with its calculated size, even though
part of it is not used. In the trace file should be written only the
compressed data block, with its real size.

Link: https://lore.kernel.org/linux-trace-devel/20220411041038.132453-1-tz.stoyanov@gmail.com

fixes: 3f8447b1 ("trace-cmd library: Add support for compression algorithms")
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>