trace-cmd: Move vsocket code into its own file
Create trace-vsock.c to hold vsock specific operations. Several files
duplicate the implementation of vsocket connections. Consolidate all of
them into a single file to be used by all, and also can be disabled in
a single place, getting rid of a lot of the #ifdef VSOCKs that are
poluting the code.

Link: https://lore.kernel.org/linux-trace-devel/20220417182154.1041513-8-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>