trace-cmd: Move add_guest_info() into trace-vm.c
In preparation for the agent proxy, move add_guest_info() into trace-vm.c
and rename it to trace_add_guest_info(). This will be used by the agent
when acting as a proxy.

Also comment the function to state what it does and what it records into
the trace.dat option.

Link: https://lore.kernel.org/linux-trace-devel/20220514024756.1319681-3-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>