trace-cmd record: Add trace_tsync_as_host() helper
In the process of adding an agent proxy, it will be useful to have a
function that will allow the agent to synchronize as the host. Move the
code to do so into a helper function that can be used by the agent. This
will be later moved to a separate file.

Link: https://lore.kernel.org/linux-trace-devel/20220514024756.1319681-16-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>