trace-cmd msg: Add PROXY communication
Add communication for an agent proxy. When an agent is also running as a
proxy, it will be able to receive a TRACE_PROXY message. This will contain
the same data as a TRACE_REQ, but also include the number of CPUs the
recorder has (as the recorder may be running on the guest, and the agent
on the host), as well as the number of siblings guests that will be sent
afterward.

The sibling messages will come at a later time.

Currently there are no users of this API.

Link: https://lore.kernel.org/linux-trace-devel/20220514024756.1319681-18-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>