trace-cmd agent proxy: Allow agent to send more meta data after trace
The options of the agent proxy trace file will need to include timings
and guest data that is added after the trace has completed. Do not close
the network handle for the meta data before the trace. Keep it open for
the proxy agent and finish writing the options at the end of the trace.

Link: https://lore.kernel.org/linux-trace-devel/20220514024756.1319681-23-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>