trace-cmd time-sync: Pass role to tracecmd_tsync_get_proto()
If the agent is a proxy server (host), then it should not pick the kvm
protocol if it can not find the kvm offset and multiplier. If it does, it
will fail the sync on the agent side which locks up the recorder side
which will still run.

This also shows how fragile the tsync code is but that is another project
to have fixed.

Link: https://lore.kernel.org/linux-trace-devel/20220519204328.534a1f30@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>