trace-cmd listen: Use copy of host for connect_ip()
If an old trace-cmd only uses v1 in the listen communication, to determine
this a trick is done by passing the v1 CPU communication a -1 followed by
the communication version. V3 and beyond knows this as the version
communication number, but v1 will error out. Then the connection is closed
and v1 is tried anew. (Note, v2 never existed in a release so it can be ignored).

The problem is that on the retry, the host has been updated with via a
strchr(), and modified. The second time around, it will not have the port
number attached to it, so it will fail to communicate.

Link: https://lore.kernel.org/linux-trace-devel/20220521180844.2af85772@gandalf.local.home

Fixes: 194467199c076 ("trace-cmd listen: Add vsocket usage")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>