trace-cmd stat: Show all filter functions that are enabled
The way the files are printed in trace-cmd stat uses a read loop and keeps
allocating memory to read the content. It reads at BUFSIZ increments. But
if a read does not read exactly BUFSIZ it stops reading and returns what
it read but may not return the rest.

This does not work with the filter functions as the read may not fill the
buffer (it does not return an unfinished function). This makes the loop
exit early.

Instead just increase with BUFSIZ against what was allocated and comparing
to the total.

The way to trigger the issue is by:

   # trace-cmd start -p nop -l "*"
   # trace-cmd show --ftrace_filter |wc -l
 50979
   # trace-cmd stat | wc -l
 234

Link: https://lore.kernel.org/linux-trace-devel/20220607183142.00dac63f@gandalf.local.home

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=212489
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>