trace-cmd library: Use uncompressed size for file size
For tracecmd_read_at() to work, each offset for every record in the file
needs to be unique. To do so, each CPU's file_offset is from the last
CPU's file_offset plus its size. But the size used was the compressed size
and not the uncompressed size.

This had "interesting" effects, where some lookups were failing. It also
caused tracecmd_read_cpu_last() to not return the last event.

Link: https://lore.kernel.org/linux-trace-devel/20220608105243.7eef6394@gandalf.local.home

Fixes: add83e0c8b511 ("trace-cmd library: Fix tracecmd_read_at()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>