trace-cmd: Fix man page examples formatting
The man page example code is only surrounded by dashes, which worked in
the early days but now it shows:

	.ft c

around the example if not the right packages are installed. Add the
"[source,<lang>]" attribute around them, and shorten the dashes to just
four. This allows the parsing to be better and the man page to be proper.

Link: https://lore.kernel.org/linux-trace-devel/20220608124832.12526bc3@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>