trace-cmd library: Fix reading of temp file used to decompress
When not using read_zpage, and reading from a temporary file, the offsets
are not accounted for properly. That is, it uses the offset of the
location of the trace.dat file to map the temporary file (that starts at
offset zero).

Account for mapping of the temporary file when appropriate.

Link: https://lore.kernel.org/linux-trace-devel/20220615025436.568682-3-rostedt@goodmis.org

Reported-by: Johannes Berg <johannes@sipsolutions.net>
Fixes: add83e0c8b51 ("trace-cmd library: Fix tracecmd_read_at()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>