trace-cmd: Move clock_context_init() out of pthreads
If the clock_context_init() fails, it can cause the communications between
the host and guest to hang without any clue to why it happened. I spent
several hours debugging this.

There's no reason that the clock_context_init() needs to be called in the
pthread. Do that before creating the threads, and move the proto into the
tsync structure itself.

Now when it fails, more can be known at the time it happens.

Link: https://lore.kernel.org/linux-trace-devel/20220707190726.3f1c6b4d@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>