trace-cmd: Have tracecmd_write_guest_time_shift() handle NULL tsync
If the clock_context_init() fails, the ctx->instance->tsync will be NULL,
but it may be used in calling tracecmd_write_guest_time_shift(). This will
cause a SEGFAULT as that function expects it to be non NULL.

Test for tsync not NULL and return -1 if it is.

More should be done to warn when this happens, but for now prevent the
SEGFAULT. This makes the code more robust.

Link: https://lore.kernel.org/linux-trace-devel/20220707191621.77deba71@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>