trace-cmd: Do not use instance from trace context
When trace-cmd initiates a connection to a trace agent over the network,
the logic in connect_to_agent() function incorrectly uses the last
instance saved in the trace context, instead of the actual instance
which is passed as input argument. This works if the remote agent is
set last on the command line, but causes a problem if there is more than
one agent or if there is a local buffer after the agent on the command
line.

Link: https://lore.kernel.org/linux-trace-devel/20220729040116.175015-1-tz.stoyanov@gmail.com

Reported-by: Alexander Aring <aahringo@redhat.com>
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Tested-by: Alexander Aring <aahringo@redhat.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>